Implement naive version of 'cycle' as in paper
This implementation is more closely related to the paper definition of
'cycle'. The initial implementation was an improvement on the generation
of the logic program (but not necessarely on the Datalog engine side)
because it generate far fewer new constants.

We decided to stick with the naive version to be as close as possible to
the paper description.

This temporarely breaks some tests.