Generalize dependency graph generation
The code to generate the dependency graph has been moved in the
companion object of the generic OWL 2 ontology wrapper Ontology.
This signals that we could potentially build a dependency graph for any
ontology (and not only RSA ontology).

Moreover, a dependency graph can be build for an Ontology object or an
arbitrary TBox and Abox.